requirements.txt
# Core Libraries
numpy
pandas
matplotlib
seaborn
scipy
pyyaml>=6.0

# Machine Learning and Data Mining
scikit-learn

# Deep Learning (PyTorch)
torch
torchvision
pillow

# Transfer Learning & SOTA Models
timm>=0.9.0
transformers>=4.30.0

# Explainability & Interpretability
grad-cam>=1.4.0
shap>=0.42.0

# Jupyter Notebooks for Analysis
jupyter
ipykernel
notebook       # Ensures Jupyter Notebook functionality
ipython        # Interactive Python shell

# Progress Bar
tqdm

# Enhanced Visualization (Optional)
plotly         # For interactive plots (Optional)

# Generating summary of the models
torchinfo

# Additional Dependencies for Dimensionality Reduction and XGBoost
umap-learn
xgboost

# Data Augmentation
albumentations>=1.3.0
imgaug>=0.4.0

# MLOps & Monitoring
mlflow>=2.0.0
wandb>=0.14.0

# API & Production
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
python-multipart>=0.0.6

# Dashboard & Visualization
streamlit>=1.28.0
gradio>=4.0.0

# Monitoring & Metrics
prometheus-client>=0.17.0